Simplify schema validation
