+++ tests/unit_tests/data_model/test_data_reader.py
@@ -4,7 +4,6 @@ import json
 import logging
 
 import astropy.units as u
-import jsonschema
 import pytest
 from astropy.io.registry.base import IORegistryError
 from astropy.table import Table
@@ -108,10 +107,3 @@ def test_read_value_from_file_and_validate(
             validate=True,
         )
     assert "Successful validation of yaml/json file" in caplog.text
-
-    # no schema given
-    test_dict_1 = {"Value": 5.0}
-    with open(tmp_test_directory / JSON_TEST_FILE, "w", encoding="utf-8") as f:
-        json.dump(test_dict_1, f)
-    with pytest.raises(jsonschema.exceptions.ValidationError):
-        data_reader.read_value_from_file(tmp_test_directory / JSON_TEST_FILE, validate=True)
